feat(html reporter): move html reporter to the core package
