+++ packages/core/src/reporters/html/HtmlReporter.ts
@@ -8,7 +8,7 @@ import { mutationTestReportSchema, Reporter } from '@stryker-mutator/api/report'
 import fileUrl = require('file-url');
 
 import { bindMutationTestReport } from './templates/bindMutationTestReport';
-import * as util from './util';
+import * as HtmlReporterUtil from './HtmlReporterUtil';
 
 const DEFAULT_BASE_FOLDER = path.normalize('reports/mutation/html');
 export const RESOURCES_DIR_NAME = 'strykerResources';
@@ -33,13 +33,13 @@ export default class HtmlReporter implements Reporter {
     const indexFileName = path.resolve(this.baseDir, 'index.html');
     await this.cleanBaseFolder();
     await Promise.all([
-      util.copyFile(
+      HtmlReporterUtil.copyFile(
         require.resolve('mutation-testing-elements/dist/mutation-test-elements.js'),
         path.resolve(this.baseDir, 'mutation-test-elements.js')
       ),
-      util.copyFile(path.resolve(__dirname, 'templates', 'stryker-80x80.png'), path.resolve(this.baseDir, 'stryker-80x80.png')),
-      util.copyFile(path.resolve(__dirname, 'templates', 'index.html'), path.resolve(this.baseDir, 'index.html')),
-      util.writeFile(path.resolve(this.baseDir, 'bind-mutation-test-report.js'), bindMutationTestReport(report))
+      HtmlReporterUtil.copyFile(path.resolve(__dirname, 'templates', 'stryker-80x80.png'), path.resolve(this.baseDir, 'stryker-80x80.png')),
+      HtmlReporterUtil.copyFile(path.resolve(__dirname, 'templates', 'index.html'), path.resolve(this.baseDir, 'index.html')),
+      HtmlReporterUtil.writeFile(path.resolve(this.baseDir, 'bind-mutation-test-report.js'), bindMutationTestReport(report))
     ]);
     this.log.info(`Your report can be found at: ${fileUrl(indexFileName)}`);
   }
@@ -60,7 +60,7 @@ export default class HtmlReporter implements Reporter {
   }
 
   private async cleanBaseFolder(): Promise<void> {
-    await util.deleteDir(this.baseDir);
-    await util.mkdir(this.baseDir);
+    await HtmlReporterUtil.deleteDir(this.baseDir);
+    await HtmlReporterUtil.mkdir(this.baseDir);
   }
 }
