+++ packages/core/src/reporters/html/templates/index.html
