+++ packages/core/src/reporters/html/templates/stryker-80x80.png
