+++ packages/core/src/reporters/index.ts
@@ -6,6 +6,7 @@ import DotsReporter from './DotsReporter';
 import EventRecorderReporter from './EventRecorderReporter';
 import ProgressAppendOnlyReporter from './ProgressAppendOnlyReporter';
 import ProgressReporter from './ProgressReporter';
+import HtmlReporter from './html/HtmlReporter';
 
 export const strykerPlugins = [
   declareClassPlugin(PluginKind.Reporter, 'clear-text', ClearTextReporter),
@@ -13,5 +14,6 @@ export const strykerPlugins = [
   declareClassPlugin(PluginKind.Reporter, 'progress-append-only', ProgressAppendOnlyReporter),
   declareClassPlugin(PluginKind.Reporter, 'dots', DotsReporter),
   declareClassPlugin(PluginKind.Reporter, 'event-recorder', EventRecorderReporter),
+  declareClassPlugin(PluginKind.Reporter, 'html', HtmlReporter),
   declareFactoryPlugin(PluginKind.Reporter, 'dashboard', dashboardReporterFactory)
 ];
