+++ packages/html-reporter/package.json
@@ -4,9 +4,6 @@
   "description": "An html reporter for the JavaScript mutation testing framework Stryker",
   "main": "src/index.js",
   "scripts": {
-    "test": "nyc --exclude-after-remap=false --check-coverage --reporter=html --report-dir=reports/coverage --lines 95 --functions 95 --branches 90 npm run mocha",
-    "mocha": "mocha \"test/helpers/**/*.js\" \"test/unit/**/*.js\" && mocha --timeout 20000 \"test/helpers/**/*.js\" \"test/integration/**/*.js\"",
-    "stryker": "node ../core/bin/stryker run"
   },
   "repository": {
     "type": "git",
@@ -32,22 +29,6 @@
     "url": "https://github.com/stryker-mutator/stryker/issues"
   },
   "homepage": "https://github.com/stryker-mutator/stryker/tree/master/packages/html-reporter#readme",
-  "dependencies": {
-    "@stryker-mutator/api": "^2.5.0",
-    "@stryker-mutator/util": "^2.5.0",
-    "file-url": "~3.0.0",
-    "mkdirp": "~1.0.3",
-    "mutation-testing-elements": "^1.0.2",
-    "rimraf": "~3.0.0"
-  },
-  "peerDependencies": {
-    "@stryker-mutator/core": "^2.0.0"
-  },
-  "devDependencies": {
-    "@stryker-mutator/test-helpers": "^2.5.0",
-    "@types/jsdom": "~12.2.0",
-    "@types/node": "^10.11.5"
-  },
   "contributors": [
     "Nico Jansen <jansennico@gmail.com>",
     "Simon de Lang <simondelang@gmail.com>"
