+++ packages/html-reporter/src/index.ts
@@ -1,5 +1,3 @@
-import { declareClassPlugin, PluginKind } from '@stryker-mutator/api/plugin';
-
-import HtmlReporter from './HtmlReporter';
-
-export const strykerPlugins = [declareClassPlugin(PluginKind.Reporter, 'html', HtmlReporter)];
+console.warn(
+  'This package is a stub package. The html reporter is now included with @stryker-mutator/core. There is no need to have the @stryker-mutator/html-reporter package installed!'
+);
