--- packages/html-reporter/stryker.conf.js
@@ -1,7 +1,0 @@
-const path = require('path');
-const settings = require('../../stryker.parent.conf');
-const moduleName = __dirname.split(path.sep).pop();
-settings.dashboard.module = moduleName;
-module.exports = function (config) {
-  config.set(settings);
-};
