--- packages/html-reporter/test/helpers/initChai.ts
@@ -1,6 +1,0 @@
-import * as chai from 'chai';
-import * as chaiAsPromised from 'chai-as-promised';
-import * as sinonChai from 'sinon-chai';
-
-chai.use(chaiAsPromised);
-chai.use(sinonChai);
