--- packages/html-reporter/test/helpers/initSinon.ts
@@ -1,7 +1,0 @@
-import { testInjector } from '@stryker-mutator/test-helpers';
-import * as sinon from 'sinon';
-
-afterEach(() => {
-  sinon.restore();
-  testInjector.reset();
-});
