--- packages/html-reporter/test/helpers/initSourceMaps.ts
@@ -1,1 +1,0 @@
-import 'source-map-support/register';
